deprecated/python_agno_version/requirements.txt
agno>=0.1.0
openai>=1.0.0
python-telegram-bot>=20.0
python-dotenv>=1.0.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
motor>=3.3.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0  # Required for TestClient
aiohttp>=3.8.0
psutil>=5.9.0 